- relevant of article popularity

+++ app/Containers/Article/Data/Migrations/2021_03_04_114732_create_article_popularity_table.php
@@ -16,7 +16,7 @@ class CreateArticlePopularityTable extends Migration
 
         $table->integer('article_id')->unsigned()->index()->nullable();
         $table->bigInteger('views_count')->unsigned()->default(0);
-        $table->integer('popularity')->unsigned()->default(0);
+        $table->float('popularity')->unsigned()->default(0);
 
         $table->date('date');
